+++ src/main/java/de/redsix/dmncheck/CheckerMain.java
@@ -1,32 +1,36 @@
 package de.redsix.dmncheck;
 
 import de.redsix.dmncheck.result.PrettyPrintValidationResults;
-import de.redsix.dmncheck.result.ValidationResult;
 import de.redsix.dmncheck.result.Severity;
+import de.redsix.dmncheck.result.ValidationResult;
+import de.redsix.dmncheck.util.ProjectClassLoader;
 import de.redsix.dmncheck.validators.core.GenericValidator;
 import de.redsix.dmncheck.validators.core.SimpleValidator;
 import de.redsix.dmncheck.validators.core.Validator;
 import io.github.lukehutch.fastclasspathscanner.FastClasspathScanner;
+import org.apache.maven.artifact.Artifact;
 import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
+import org.apache.maven.plugins.annotations.ResolutionScope;
+import org.apache.maven.project.MavenProject;
 import org.camunda.bpm.model.dmn.Dmn;
 import org.camunda.bpm.model.dmn.DmnModelInstance;
 
 import java.io.File;
 import java.io.IOException;
 import java.lang.reflect.Modifier;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.net.URLClassLoader;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
+import java.util.*;
 import java.util.stream.Collectors;
 
-@Mojo(name = "check-dmn", requiresProject = false)
+@Mojo(name = "check-dmn", requiresProject = false, requiresDependencyResolution = ResolutionScope.TEST)
 class CheckerMain extends AbstractMojo {
 
     private static final String VALIDATOR_PACKAGE = "de.redsix.dmncheck.validators";
@@ -43,8 +47,14 @@ class CheckerMain extends AbstractMojo {
     @SuppressWarnings("nullness")
     private String[] validators;
 
+    @Parameter( defaultValue = "${project}", readonly = true )
+    @SuppressWarnings("nullness")
+    private MavenProject project;
+
     @Override
     public void execute() throws MojoExecutionException {
+        loadProjectclasspath();
+
         final List<Path> searchPathObjects = getSearchPathList().stream().map(Paths::get).collect(Collectors.toList());
         final List<File> filesToTest = fetchFilesToTestFromSearchPaths(searchPathObjects);
 
@@ -142,7 +152,6 @@ class CheckerMain extends AbstractMojo {
         final List<Class<? extends Validator>> validatorClasses = new ArrayList<>();
         new FastClasspathScanner(scanSpec)
                 .disableRecursiveScanning()
-                .strictWhitelist()
                 .matchSubclassesOf(GenericValidator.class, validatorClasses::add)
                 .matchSubclassesOf(SimpleValidator.class, validatorClasses::add)
                 .matchClassesImplementing(Validator.class, validatorClasses::add)
@@ -163,6 +172,24 @@ class CheckerMain extends AbstractMojo {
         }
     }
 
+    private void loadProjectclasspath() throws MojoExecutionException {
+        final List<URL> listUrl = new ArrayList<>();
+
+        Set<Artifact> deps = project.getArtifacts();
+        for (Artifact artifact : deps) {
+            final URL url;
+            try {
+                url = artifact.getFile().toURI().toURL();
+                listUrl.add(url);
+            }
+            catch (MalformedURLException e) {
+                throw new MojoExecutionException("Failed to construct project class loader.");
+            }
+        }
+
+        ProjectClassLoader.instance.classLoader = new URLClassLoader(listUrl.toArray(new URL[0]));
+    }
+
     void setExcludes(final String[] excludes) {
         this.excludes = excludes;
     }
@@ -175,4 +202,7 @@ class CheckerMain extends AbstractMojo {
         this.validators = validators;
     }
 
+    public void setProject(MavenProject project) {
+        this.project = project;
+    }
 }