+++ src/main/java/de/redsix/dmncheck/validators/TypeValidator.java
@@ -1,26 +1,27 @@
 package de.redsix.dmncheck.validators;
 
+import de.redsix.dmncheck.feel.ExpressionTypes;
 import de.redsix.dmncheck.feel.FeelParser;
 import de.redsix.dmncheck.feel.FeelTypecheck;
 import de.redsix.dmncheck.feel.ExpressionType;
 import de.redsix.dmncheck.result.ValidationResult;
-import de.redsix.dmncheck.util.Either;
-import de.redsix.dmncheck.util.Eithers;
+import de.redsix.dmncheck.util.ProjectClassLoader;
 import de.redsix.dmncheck.util.Util;
 import de.redsix.dmncheck.validators.core.SimpleValidator;
 import org.camunda.bpm.model.dmn.instance.DecisionTable;
 import org.camunda.bpm.model.dmn.instance.DmnElement;
 import org.camunda.bpm.model.dmn.instance.Rule;
 
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 public abstract class TypeValidator extends SimpleValidator<DecisionTable> {
 
     abstract String errorMessage();
-
-    abstract boolean isEmptyAllowed();
 
     Stream<ValidationResult> typecheck(final Rule rule, final Stream<? extends DmnElement> expressions, final Stream<String> variables,
             final Stream<ExpressionType> types) {
@@ -44,16 +45,36 @@ public abstract class TypeValidator extends SimpleValidator<DecisionTable> {
 
     private List<ValidationResult.Builder.BuildStep> typecheckExpression(Rule rule, DmnElement inputEntry, FeelTypecheck.Context context,
             ExpressionType expectedType) {
-        final Either<ExpressionType, ValidationResult.Builder.ElementStep> typedcheckResult = FeelParser.parse(inputEntry.getTextContent())
-                .bind(feelExpression -> FeelTypecheck.typecheck(context, feelExpression));
+        return FeelParser.parse(inputEntry.getTextContent()).bind(feelExpression -> FeelTypecheck.typecheck(context, feelExpression))
+                .map(type -> {
+                    if (type.isSubtypeOf(ExpressionTypes.STRING()) && ExpressionTypes.getClassName(expectedType).isPresent()) {
+                        return checkEnumValue(ExpressionTypes.getClassName(expectedType).get(), inputEntry.getTextContent(), rule);
+                    } else if (type.isSubtypeOf(expectedType) || ExpressionTypes.TOP().equals(type)) {
+                        return Collections.<ValidationResult.Builder.BuildStep>emptyList();
+                    } else {
+                        return Collections.singletonList(ValidationResult.init.message(errorMessage()).element(rule));
+                    }
+                }).match(Function.identity(), validationResultBuilder -> Collections.singletonList(validationResultBuilder.element(rule)));
+    }
 
-        return Eithers.caseOf(typedcheckResult).makeLeft(type -> {
-            if (type.isSubtypeOf(expectedType) || isEmptyAllowed() && ExpressionType.TOP.equals(type)) {
-                return Collections.<ValidationResult.Builder.BuildStep>emptyList();
+    private List<ValidationResult.Builder.BuildStep> checkEnumValue(final String className, final String stringValue, final Rule rule) {
+        try {
+            Class<?> anEnum = ProjectClassLoader.instance.classLoader.loadClass(className);
+            if (anEnum.isEnum()) {
+                final List<String> foo = Arrays.stream(((Class<? extends Enum>) anEnum).getEnumConstants()).map(Enum::name).collect(Collectors.toList());
+                final String value = stringValue.substring(1, stringValue.length() -1 );
+                if (foo.contains(value)) {
+                    return Collections.emptyList();
+                } else {
+                    return Collections.singletonList(ValidationResult.init.message("Value " + stringValue + " does not belong to " + className).element(rule));
+                }
             } else {
-                return Collections.singletonList(ValidationResult.Builder.init.message(errorMessage()).element(rule));
+                return Collections.singletonList(ValidationResult.init.message("Class " + className + " is no enum.").element(rule));
             }
-        }).makeRight(validationResultBuilder -> Collections.singletonList(validationResultBuilder.element(rule)));
+        }
+        catch (ClassNotFoundException e) {
+            return Collections.singletonList(ValidationResult.init.message("Class " + className + " not found on project classpath.").element(rule));
+        }
     }
 
     @Override