+++ src/test/java/de/redsix/dmncheck/validators/AggregationOutputTypeValidatorTest.java
@@ -47,9 +47,28 @@ class AggregationOutputTypeValidatorTest extends WithDecisionTable {
         assertEquals(1, validationResults.size());
         final ValidationResult validationResult = validationResults.get(0);
         assertAll(
-                () -> assertEquals("An aggregation is used but no output severity is defined", validationResult.getMessage()),
+                () -> assertEquals("An aggregation is used but no output type is defined", validationResult.getMessage()),
                 () -> assertEquals(output, validationResult.getElement()),
                 () -> assertEquals(Severity.WARNING, validationResult.getSeverity())
+        );
+    }
+
+    @Test
+    void shouldErrorOnUnsupportedOutputType() {
+        decisionTable.setHitPolicy(HitPolicy.COLLECT);
+        decisionTable.setAggregation(BuiltinAggregator.MAX);
+        final Output output = modelInstance.newInstance(Output.class);
+        output.setTypeRef("unsupportedType");
+        decisionTable.getOutputs().add(output);
+
+        final List<ValidationResult> validationResults = testee.apply(modelInstance);
+
+        assertEquals(1, validationResults.size());
+        final ValidationResult validationResult = validationResults.get(0);
+        assertAll(
+                () -> assertEquals("Could not parse FEEL expression type 'unsupportedType'", validationResult.getMessage()),
+                () -> assertEquals(output, validationResult.getElement()),
+                () -> assertEquals(Severity.ERROR, validationResult.getSeverity())
         );
     }