+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+### 1.0.3 (2018-06-12)
+
+Bugfixes:
+
+  - Fix detection of shadowed rules with strings
+
+## 1.0.2 (2018-05-31)
+
+Bugfixes:
+
+  - Allow empty values in output entries
+
+Features:
+
+  - New Validator for Enum constants
+  - Warn if an aggregation is used without a type
+  - Sort validation results descinding by severity
+  - Improved error messages
+
 ## 1.0.1 (2018-04-19)
 
 Bufixes: